pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.0",
    "uvicorn>=0.41.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "python-dotenv>=1.0",
    "apscheduler>=3.10",
    "numpy>=1.24.0",
    "dimod>=0.12.21",
    "dwave-samplers>=1.4.0",
]
